refactor(agent): use dynamic import for subcommand
